requirements.txt
langchain-community>=0.0.13
langchain-google-genai>=0.0.6
langchain>=0.1.0
python-dotenv>=1.0.0
chromadb>=0.4.0
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0